+++ Sources/ParseSwift/Types/Query+async.swift
@@ -28,7 +28,7 @@ public extension Query {
 
     /**
      Query plan information for finding objects *asynchronously*.
-     - parameter isUsingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
+     - parameter usingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - note: An explain query will have many different underlying types. Since Swift is a strongly
      typed language, a developer should specify the type expected to be decoded which will be
@@ -37,13 +37,13 @@ public extension Query {
      - returns: An array of ParseObjects.
      - throws: An error of type `ParseError`.
      - warning: MongoDB's **explain** does not conform to the traditional Parse Server response, so the
-     `isUsingMongoDB` flag needs to be set for MongoDB users. See more
+     `usingMongoDB` flag needs to be set for MongoDB users. See more
      [here](https://github.com/parse-community/parse-server/pull/7440).
     */
-    func findExplain<U: Decodable>(isUsingMongoDB: Bool = false,
+    func findExplain<U: Decodable>(usingMongoDB: Bool = false,
                                    options: API.Options = []) async throws -> [U] {
         try await withCheckedThrowingContinuation { continuation in
-            self.findExplain(isUsingMongoDB: isUsingMongoDB,
+            self.findExplain(usingMongoDB: usingMongoDB,
                              options: options,
                              completion: continuation.resume)
         }
@@ -86,18 +86,18 @@ public extension Query {
      typed language, a developer should specify the type expected to be decoded which will be
      different for MongoDB and PostgreSQL. One way around this is to use a type-erased wrapper
      such as the [AnyCodable](https://github.com/Flight-School/AnyCodable) package.
-     - parameter isUsingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
+     - parameter usingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: An array of ParseObjects.
      - throws: An error of type `ParseError`.
      - warning: MongoDB's **explain** does not conform to the traditional Parse Server response, so the
-     `isUsingMongoDB` flag needs to be set for MongoDB users. See more
+     `usingMongoDB` flag needs to be set for MongoDB users. See more
      [here](https://github.com/parse-community/parse-server/pull/7440).
     */
-    func firstExplain<U: Decodable>(isUsingMongoDB: Bool = false,
+    func firstExplain<U: Decodable>(usingMongoDB: Bool = false,
                                     options: API.Options = []) async throws -> U {
         try await withCheckedThrowingContinuation { continuation in
-            self.firstExplain(isUsingMongoDB: isUsingMongoDB,
+            self.firstExplain(usingMongoDB: usingMongoDB,
                               options: options,
                               completion: continuation.resume)
         }
@@ -122,18 +122,18 @@ public extension Query {
      typed language, a developer should specify the type expected to be decoded which will be
      different for MongoDB and PostgreSQL. One way around this is to use a type-erased wrapper
      such as the [AnyCodable](https://github.com/Flight-School/AnyCodable) package.
-     - parameter isUsingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
+     - parameter usingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: An array of ParseObjects.
      - throws: An error of type `ParseError`.
      - warning: MongoDB's **explain** does not conform to the traditional Parse Server response, so the
-     `isUsingMongoDB` flag needs to be set for MongoDB users. See more
+     `usingMongoDB` flag needs to be set for MongoDB users. See more
      [here](https://github.com/parse-community/parse-server/pull/7440).
     */
-    func countExplain<U: Decodable>(isUsingMongoDB: Bool = false,
+    func countExplain<U: Decodable>(usingMongoDB: Bool = false,
                                     options: API.Options = []) async throws -> [U] {
         try await withCheckedThrowingContinuation { continuation in
-            self.countExplain(isUsingMongoDB: isUsingMongoDB,
+            self.countExplain(usingMongoDB: usingMongoDB,
                               options: options,
                               completion: continuation.resume)
         }
@@ -159,18 +159,18 @@ public extension Query {
      typed language, a developer should specify the type expected to be decoded which will be
      different for mongoDB and PostgreSQL. One way around this is to use a type-erased wrapper
      such as the [AnyCodable](https://github.com/Flight-School/AnyCodable) package.
-     - parameter isUsingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
+     - parameter usingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: An array of ParseObjects.
      - throws: An error of type `ParseError`.
      - warning: MongoDB's **explain** does not conform to the traditional Parse Server response, so the
-     `isUsingMongoDB` flag needs to be set for MongoDB users. See more
+     `usingMongoDB` flag needs to be set for MongoDB users. See more
      [here](https://github.com/parse-community/parse-server/pull/7440).
     */
-    func withCountExplain<U: Decodable>(isUsingMongoDB: Bool = false,
+    func withCountExplain<U: Decodable>(usingMongoDB: Bool = false,
                                         options: API.Options = []) async throws -> [U] {
         try await withCheckedThrowingContinuation { continuation in
-            self.withCountExplain(isUsingMongoDB: isUsingMongoDB,
+            self.withCountExplain(usingMongoDB: usingMongoDB,
                                   options: options,
                                   completion: continuation.resume)
         }
@@ -201,20 +201,20 @@ public extension Query {
      different for MongoDB and PostgreSQL. One way around this is to use a type-erased wrapper
      such as the [AnyCodable](https://github.com/Flight-School/AnyCodable) package.
      - parameter pipeline: A pipeline of stages to process query.
-     - parameter isUsingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
+     - parameter usingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: An array of ParseObjects.
      - throws: An error of type `ParseError`.
      - warning: MongoDB's **explain** does not conform to the traditional Parse Server response, so the
-     `isUsingMongoDB` flag needs to be set for MongoDB users. See more
+     `usingMongoDB` flag needs to be set for MongoDB users. See more
      [here](https://github.com/parse-community/parse-server/pull/7440).
     */
     func aggregateExplain<U: Decodable>(_ pipeline: [[String: Encodable]],
-                                        isUsingMongoDB: Bool = false,
+                                        usingMongoDB: Bool = false,
                                         options: API.Options = []) async throws -> [U] {
         try await withCheckedThrowingContinuation { continuation in
             self.aggregateExplain(pipeline,
-                                  isUsingMongoDB: isUsingMongoDB,
+                                  usingMongoDB: usingMongoDB,
                                   options: options,
                                   completion: continuation.resume)
         }
@@ -245,20 +245,20 @@ public extension Query {
      different for MongoDB and PostgreSQL. One way around this is to use a type-erased wrapper
      such as the [AnyCodable](https://github.com/Flight-School/AnyCodable) package.
      - parameter key: A field to find distinct values.
-     - parameter isUsingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
+     - parameter usingMongoDB: Set to **true** if your Parse Server uses MongoDB. Defaults to **false**.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: An array of ParseObjects.
      - throws: An error of type `ParseError`.
      - warning: MongoDB's **explain** does not conform to the traditional Parse Server response, so the
-     `isUsingMongoDB` flag needs to be set for MongoDB users. See more
+     `usingMongoDB` flag needs to be set for MongoDB users. See more
      [here](https://github.com/parse-community/parse-server/pull/7440).
     */
     func distinctExplain<U: Decodable>(_ key: String,
-                                       isUsingMongoDB: Bool = false,
+                                       usingMongoDB: Bool = false,
                                        options: API.Options = []) async throws -> [U] {
         try await withCheckedThrowingContinuation { continuation in
             self.distinctExplain(key,
-                                 isUsingMongoDB: isUsingMongoDB,
+                                 usingMongoDB: usingMongoDB,
                                  options: options,
                                  completion: continuation.resume)
         }